refactor: improve documentation and remove unused parseGnomeShellVersionString function
